refactor: game.js -> gameUtils.js
